bug fixed in oop lab03T03

+++ 2-1/oop/lab03/115_Task_03.cpp
@@ -14,7 +14,7 @@ public:
     void assignName(const char *Name) {         
         strcpy(name, Name);
     }
-    void setPrice(int price) {
+    void setPrice(double price) {
         if(price >= 0) {
             unitPrice = price;
         } else cout << "Invalid price\n";
